requirements.txt
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
outcome==1.3.0.post0
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-telegram-bot==20.8
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
telebot==0.0.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==1.26.6
wsproto==1.2.0